Updates to the getting started docs
Call the Getting Started guide a tutorial since that's the style it's written in

Update the information about the phoenix guide

Upgrade ex_doc to the latest (improves some formatting and a bug I
encountered related to commented code)

Make more links internal instead of going through hexdocs (this will
keep them more up-to-date as new versions get created)

Update the ``# attribute :created_at, :utc_datetime_usec` to
`:utc_datetime_usec` to match the note in the README

Various language changes

Update the DataLayer table to have more descriptive names and combine
the "Documentation" column with the "DataLayer" column.

More clearly mark the ETS DataLayer as being for testing (to match it's
own moduledoc)

Remove note about ETS only being read/writeable by the owning process
because it's not clear if that's a default of Ash.DataLayer.Ets and it
doesn't seem relevant for the tutorial.