docs: did a rework of initial get-started-with-phoenix draft
I was copying the helpdesk example from the pure ash getting started article.
But there was too much extra content which made the guide very long.
Instead I decided the make the resource very minimal, included lots of
information about how to get ash up and running with phoenix and ash postgres.

I'm very welcome to recieve feedback on what I've done so far, I want it to be
as good as it can be.