docs: introduces new attributes topic
The attribute docs are really quite good. There's not much to cover apart from
constraints. But constraints also apply to action arguments too so they don't
really belong here.

One area where the attribute docs are lacking is the explaination of some of the
options. I was thinking about covering a few of them in the attributes topic
but I think its probably best to beef up the options section in the attribute docs.