docs: initial draft of standalone constraints article
I'm not sure how I feel about it being standalone. I think there's enough
content for it to be standalone given the size of some of the other topics.
Also the only other logical place to put it is `Ash.Type` but it doesn't feel
like it _quite_ fits in there to me.