improvement: better ergonomics for atomic updates
We need to generally improve ergonomics about working with atomic updates
generically. Specifically, any code that wants to tap into something changing
has to be aware that it could be changing atomically or not. This change
makes it so that *only* expression updates are written to the atomics set.
This means if `changeset.atomics[key]` exists, then it is an expression or
something that otherwise cannot be manipulated.