improvement: allow data loading when no primary read action exists
this is still a bit experimental, so there may be some sharp edges,
but it makes sense for a resource with only a create action to be able to
load data, but to do so it currently needs to add a read action that will
never actually be called, only "used" as the action being performed, for things
like policies etc. We may even want to make this logic the default in a future
major version.