improvement: default backwards compatible interface to false for domains
in 2.x, we had `YourDomain.read` and `YourDomain.create`, etc.
For the sake of backards compatibility, we contineud to define those functions.
However, it has been deprecated for long enough now that it should be safe to
change the default. If you need those functions, do:

`use Ash.Domain, backwards_compatible_interface?: true`