improvement: don't show required errors for fields with other errors
a very common behavior is to have a change that will set an attribute or
argument that would otherwise be required, or set some error. In
situations like that, you could have an error saying "x is required" and
another error saying something like "x could not be set because of some
reason". This is generally speaking a bad experience, and for that
reason we automatically hide required errors if there are other errors
for the field that is required