fix: don't allow modifying changeset with atomic conditions
this fixes a bug where by some conditions on changes were not being
honored. This may cause a problem for certiain specific formulations
of changes plus conditions right now, but the current behavior cannot
be kept. We will likely need a way to return after_action hooks that
will be called conditionally if the expression returns true or false
when placed as a calcualtion into the update. Not an easy change
unfortunately.

improvement: add `touching?` option to changing validation