fix: dump values to native storage type when doing atomic upgrades improvement: add compile flag requiring atomic for default actions
Ash has a layer of protection around writing update actions that are
not friendly to asynchronous execution, called `require_atomic?`. This
was *supposed* to default to `true` for all update actions, but it was
not doing so for the actions defined with `defaults: [update: [...]]`.

New applications will enforce this by default, but to opt into this new
(and better) behavior, set the following config:

```elixir
config :ash, default_actions_require_atomic?: true
```

It is very important to keep in mind that *not all non-atomic actions
can be detected at compile time*!. This means that adopting this new
configuration can cause *runtime* errors. This is why we have tests.

improvement: types support atomic update by default for non-expr values

There is a *very very very* slim chance that this will be a breaking
change. However, its such a little known feature that you can even
*have* types that depend on their previous values. To my knowledge
the only place that leverages this is with embedded resources, and
those already handle this logic appropriately.