improvement: Ash.Type.Enum - Add optional description and label value … (#1925)
- Add keyword list for enum values to add `description` and/or `label`
- Add `label/1` to return the label for a given value
- Add `details/1` to return the label and description map
- Add validation to raise when unsupported options are provided for a value
- Add `value_to_label/1` function that will generate a label from the `:atom` key when no label
is provided.  For `:string` keys, the label is set to the key value.
Example

```elixir
defmodule MyApp.MyEnum do
  use Ash.Type.Enum, values: [
    open: "An open ticket",
    closed: [description: "A closed ticket", label: "Closed"],
    escalated: [label: "Uh oh!"]
  ]
end

```

This addition is completely backwards compatible and all existing tests pass.  Ash users that
already have enums with descriptions, won't be affected.

I attempted to switch away from the `validate_values` function and leverage constraints,
but didn't have much luck.  Open to guidance on how to do it as I think it would be nice
to do instead of the manual validation function.

closes #1724