improvement: add `Ash.Scope` for converting a scope to options
This would likely need to be incrementally adopted in various places,
and the first place would be in `AshAuthentication`. We would likely
make it an option in the `ash_authentication_live_session`, to take a
`scope`, and expect that scope to define a function like `from_conn/1`.
(or something like that)