fix: prevent MustBeAtomic error on embedded attribute updates
There are 3 documented conditions for atomic updates to work on embedded resources:

> Reason: Embedded attributes do not support atomic updates unless they have no primary key, or `constraints[:on_update]` is set to `:replace`, or the update action accepts all public attributes and has no changes.

But when an embedded attribute is not an array, the atomic action check was failing it.

This was discussed on [ElixirForum](https://elixirforum.com/t/atomic-updates-across-embedded-and-normal-attributes-failing-since-3-5-10/70936) where Zach suggested this fix. I'm PR'ing it to try and be helpful.

I couldn't find an appropriate test to modify. There is existing coverage of a `Profile` resource in `embedded_resource_test.exs`, but this is currently set to not require atomic updates. If I require atomic updates and set `on_update: :replace`, the existing assertions around validations are broken. I didn't want to create an entire duplicate Profile test to cover this small case.