docs: Update before_action docs to use force_change_attribute/2 (#2258)
Using change_attribute/2 in a before_action hook triggers a warning
that the changeset has already been validated, suggests a new pattern
that isn't compatible with the before_action hook. This confused me at
first when copying the patterns from the documentation, so I suggest
using force_change_attribute in the examples instead, as well as adding
a small note about why.