fix: Ash.Query.after_transaction result argument issues. (#2354)
* fix: after_transaction on read error bug.

Fixes an issue where when an `after_action` callback on a read query failing would cause the `after_transaction` hook to receive a tuple of the error and the query as it's `result` argument instead of just the error.

* fix: correct after_transaction result type for successful queries.

Previously it was passing a large ok tuple to the after transaction hook instead of just the `{:ok, results}` result that `t:Ash.Query.after_transaction_fun/0` would suggest.