fix: Trim triple newlines instead of double newlines
Removing double newlines breaks formatting in e.g. the generated browser docs,
and breaks doctests, since they expect the example code sections to be separated
by the rest of the comments by double newlines.

Instead, we should focus on collapsing triple newlines, which can occur for
legitimate reasons, e.g. a section of the generated documentation doesn't return
any text, but which shouldn't be showed in the resulting documentation.

Fixes: #2433